Remove default size in nids::vector
Remove the default allocation size of nids::vector (previously 8)
so that new objects allocated with a nids::vector do not have an
incidental cost associated with them.

Modify nids::vector so that it accepts 0 capacity so that a nids::vector
can be initialized to 0, and so that the resize methods can handle that
as an input from the push_back methods.