Fix destructor, fix bug in vector::shrink_to_fit
The destructor for nids::vector now calls the destructor for the items
it contains, which should have been the default behavior.

Fixed a bug in shrink_to_fit which caused m_array to no longer point at
the start of the allocated space, causing free to fail in the destructor
in release builds.